[Silver III] Title: 주유소, Time: 92 ms, Memory: 3588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -43,8 +43,8 @@ int main()
 	
 	*/
 	
-	int cur = cost[0];
-	int capdis = 0;
+	ll cur = cost[0];
+	ll capdis = 0;
 	for (int i = 0; i < n; ++i)
 	{
 		if (cur > cost[i] || i == n-1)
